+++ src/index.ts
@@ -131,6 +131,7 @@ import * as UnitPriceOfferFactory from './unitPriceOffer';
 
 import * as TaskFactory from './task';
 import * as AccountMoneyTransferTaskFactory from './task/accountMoneyTransfer';
+import * as AggregateOffersTaskFactory from './task/aggregateOffers';
 import * as AggregateScreeningEventTaskFactory from './task/aggregateScreeningEvent';
 import * as AggregateUseActionsOnEventTaskFactory from './task/aggregateUseActionsOnEvent';
 import * as CancelAccountMoneyTransferTaskFactory from './task/cancelAccountMoneyTransfer';
@@ -586,7 +587,6 @@ export namespace task {
         T extends TaskName.VoidRegisterServiceTransaction ? VoidRegisterServiceTransactionTaskFactory.IData :
         T extends TaskName.VoidReserveTransaction ? VoidReserveTransactionTaskFactory.IData :
         TaskFactory.IData;
-
     export type IAttributes<T extends TaskName | string> =
         T extends TaskName.ConfirmReserveTransaction ? ConfirmReserveTransactionTaskFactory.IAttributes :
         T extends TaskName.CreateEvent ? CreateEventTaskFactory.IAttributes :
@@ -617,7 +617,6 @@ export namespace task {
         T extends TaskName.VoidRegisterServiceTransaction ? VoidRegisterServiceTransactionTaskFactory.IAttributes :
         T extends TaskName.VoidReserveTransaction ? VoidReserveTransactionTaskFactory.IAttributes :
         TaskFactory.IAttributes;
-
     export type ITask<T extends TaskName | string> =
         T extends TaskName.ConfirmReserveTransaction ? ConfirmReserveTransactionTaskFactory.ITask :
         T extends TaskName.CreateEvent ? CreateEventTaskFactory.ITask :
@@ -648,11 +647,10 @@ export namespace task {
         T extends TaskName.VoidRegisterServiceTransaction ? VoidRegisterServiceTransactionTaskFactory.ITask :
         T extends TaskName.VoidReserveTransaction ? VoidReserveTransactionTaskFactory.ITask :
         TaskFactory.ITask;
-
     export import ISearchConditions = TaskFactory.ISearchConditions;
-
     export import IExecutionResult = TaskFactory.IExecutionResult;
 
+    export import aggregateOffers = AggregateOffersTaskFactory;
     export import aggregateScreeningEvent = AggregateScreeningEventTaskFactory;
     export import aggregateUseActionsOnEvent = AggregateUseActionsOnEventTaskFactory;
     export import cancelMoneyTransfer = CancelMoneyTransferTaskFactory;