+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- 経理レポートの返金アクションをFailedActionStatusに対応
+
 ### Deprecated
 
 ### Removed