+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - 経理レポートの返金アクションをFailedActionStatusに対応
 - 決済承認アクションのobjectを最適化
+- イベントの予約集計IFを最適化
 
 ### Deprecated